refactor(tests): move get_bound_llm_magic_mock to utils
